// src/lib/constants/action-types.ts
export enum FETCH_ACTIONS {
    FETCH_START = 'FETCH_START',
    FETCH_SUCCESS = 'FETCH_SUCCESS',
    FETCH_ERROR = 'FETCH_ERROR',
};

export enum ERROR_ACTIONS {
    CLEAR_ERROR = 'CLEAR_ERROR',
    CLEAR_ERRORS = 'CLEAR_ERRORS',
};

export enum GEOCODING_ACTIONS {
    CLEAR_COORDS = 'CLEAR_COORDS',
};

export enum WIDGET_ACTIONS {
    SET_APPID = 'SET_APPID',
    SET_SHOW_WIDGET_FIELD = 'SET_SHOW_WIDGET_FIELD',
    SET_FLAVOR = 'SET_FLAVOR',
    SET_DEFAULT_CITY = 'SET_DEFAULT_CITY',
};

export enum TEST_ACTIONS {
    TEST_ACTION = 'TEST_ACTION'
};